requirements.txt.txt
flask
gunicorn
flask-mysql
